Move PulseAnimation to :browser-ui
Relocate PulseAnimation (and its ic_circle_pulse_blue drawable) from :app to :browser-ui so it can be shared with duckchat-impl. Pure move with no behavior change; updates the three import sites (OmnibarLayout, BrowserNavigationBarView, NativeInputManager).
