+++ browser/browser-ui/src/main/res/drawable/ic_circle_pulse_blue.xml
